+++ addOns/zest/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Print statements should print to the relevant script Output tab.
+- Recording client Zest scripts.
 
 ## [48.5.0] - 2025-03-25
 ### Changed